BODY: adde a cast in dxf_body_write().

+++ src/body.c
@@ -549,7 +549,7 @@ dxf_body_write
         }
         if (body->paperspace == DXF_PAPERSPACE)
         {
-                fprintf (fp->fp, " 67\n%hd\n", DXF_PAPERSPACE);
+                fprintf (fp->fp, " 67\n%hd\n", (int16_t) DXF_PAPERSPACE);
         }
         fprintf (fp->fp, "  8\n%s\n", body->layer);
         if (strcmp (body->linetype, DXF_DEFAULT_LINETYPE) != 0)
